Rename logging.py to log.py to fix name collision
Trying to import logging breaks because it finds that local file first.
