+++ butter/network.py
@@ -4,7 +4,7 @@ Butter Network
 This component should allow for intuitive and transparent control over networks, which are the top
 level containers for groups of instances/services.
 """
-from butter.logging import logger
+from butter.log import logger
 from butter.providers import get_provider
